# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Talent Match Backend API",
    "version": "1.0.0",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://talent-match-backend.onrender.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/teacher_info/": {
      "get": {
        "tags": [
          "Teacher_info"
        ],
        "summary": "取得老師資料",
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTeacherInfoResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/GetTeacherInfoResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/teacher_video/": {
      "get": {
        "tags": [
          "Teacher_video"
        ],
        "summary": "取得所有影片",
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTeacherVideoListResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/GetTeacherVideoListResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Teacher_video"
        ],
        "summary": "新增影片",
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddTeacherVideoResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AddTeacherVideoResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTeacherVideoRequestModel"
              }
            }
          }
        }
      }
    },
...
    "/api/favorites/": {
      "get": {
        "tags": [
          "Favorites"
        ],
        "summary": "取得收藏課程",
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error500ResponseModel"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Favorites"
        ],
        "summary": "新增收藏課程",
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddFavoriteResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AddFavoriteResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteRequestModel"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Favorites"
        ],
        "summary": "取消收藏課程",
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelFavoriteResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/CancelFavoriteResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error400ResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error500ResponseModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteRequestModel"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "schemas": {
...
      "Error400ResponseModel": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "status": "false",
          "message": "錯誤訊息"
        }
      },
...
      "Error500ResponseModel": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "status": "false",
          "message": "伺服器錯誤"
        }
      },
...
      "GetTeacherInfoResponseModel": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "avator_image": {
                "type": "string"
              },
              "main_categorys": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "sub_categorys": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "nationality": {
                "type": "string"
              },
              "introduction": {
                "type": "string"
              },
              "work_experiences": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "is_working": {
                      "type": "boolean"
                    },
                    "workplace": {
                      "type": "string"
                    },
                    "job_category": {
                      "type": "string"
                    },
                    "start_year": {
                      "type": "number"
                    },
                    "start_month": {
                      "type": "number"
                    },
                    "end_year": {
                      "type": "number"
                    },
                    "end_month": {
                      "type": "number"
                    },
                    "position": {
                      "type": "string"
                    },
                    "place": {
                      "type": "string"
                    }
                  }
                }
              },
              "learning_experience": {
                "type": "object",
                "properties": {
                  "is_in_school": {
                    "type": "boolean"
                  },
                  "degree": {
                    "type": "string"
                  },
                  "department": {
                    "type": "string"
                  },
                  "start_year": {
                    "type": "number"
                  },
                  "start_month": {
                    "type": "number"
                  },
                  "end_year": {
                    "type": "number"
                  },
                  "end_month": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  },
                  "place": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string"
                  }
                }
              },
              "teaching_certificate": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "verifying_institution": {
                      "type": "string"
                    },
                    "license_name": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "license_number": {
                      "type": "string"
                    },
                    "file": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    },
                    "subject": {
                      "type": "string"
                    }
                  }
                }
              },
              "intro_video_id": {
                "type": "string"
              }
            }
          }
        },
        "example": {
          "status": true,
          "data": {
            "avator_image": "string",
            "main_categorys": [
              "string"
            ],
            "sub_categorys": [
              "string"
            ],
            "nationality": "string",
            "introduction": "string",
            "work_experiences": [
              {
                "is_working": true,
                "workplace": "string",
                "job_category": "string",
                "start_year": 2020,
                "start_month": 1,
                "end_year": 2021,
                "end_month": 12,
                "position": "string",
                "place": "string"
              }
            ],
            "learning_experience": {
              "is_in_school": true,
              "degree": "string",
              "department": "string",
              "start_year": 2018,
              "start_month": 9,
              "end_year": 2022,
              "end_month": 6,
              "name": "string",
              "place": "string",
              "file": "string"
            },
            "teaching_certificate": [
              {
                "verifying_institution": "string",
                "license_name": "string",
                "name": "string",
                "license_number": "string",
                "file": "string",
                "category": "string",
                "subject": "string"
              }
            ],
            "intro_video_id": "string"
          }
        }
      },
...
      "FavoritesResponseModel": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "狀態"
          },
          "data": {
            "type": "object",
            "properties": {
              "favorites": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "course_id": {
                      "type": "string",
                      "description": "課程ID"
                    },
                    "rate": {
                      "type": "number",
                      "description": "評分"
                    },
                    "name": {
                      "type": "string",
                      "description": "課程名稱"
                    },
                    "content": {
                      "type": "string",
                      "description": "課程內容"
                    },
                    "main_image": {
                      "type": "string",
                      "description": "主圖片"
                    },
                    "main_category": {
                      "type": "string",
                      "description": "主類別"
                    },
                    "sub_category": {
                      "type": "string",
                      "description": "子類別"
                    },
                    "teacher_id": {
                      "type": "string",
                      "description": "教師ID"
                    },
                    "teacher_name": {
                      "type": "string",
                      "description": "教師名稱"
                    },
                    "teacher_avator": {
                      "type": "string",
                      "description": "教師頭像"
                    },
                    "price_quantity": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "price": {
                            "type": "number",
                            "description": "價格"
                          },
                          "quantity": {
                            "type": "number",
                            "description": "數量"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "example": {
          "status": true,
          "data": {
            "favorites": [
              {
                "course_id": "string",
                "rate": 0,
                "name": "string",
                "content": "string",
                "main_image": "string",
                "main_category": "string",
                "sub_category": "string",
                "teacher_id": "string",
                "teacher_name": "string",
                "teacher_avator": "string",
                "price_quantity": [
                  {
                    "_id": "string",
                    "price": 0,
                    "quantity": 0
                  }
                ]
              }
            ]
          }
        }
      },
      "FavoriteRequestModel": {
        "type": "object",
        "required": [
          "course_id"
        ],
        "properties": {
          "course_id": {
            "type": "string",
            "description": "課程ID"
          }
        },
        "example": {
          "course_id": "string"
        }
      },
      "AddFavoriteResponseModel": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "狀態"
          },
          "data": {
            "type": "object",
            "properties": {
              "faviorites": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "course_id": {
                      "type": "string",
                      "description": "課程ID"
                    },
                    "rate": {
                      "type": "number",
                      "description": "評分"
                    },
                    "name": {
                      "type": "string",
                      "description": "課程名稱"
                    },
                    "content": {
                      "type": "string",
                      "description": "課程內容"
                    },
                    "main_image": {
                      "type": "string",
                      "description": "主圖片"
                    },
                    "main_category": {
                      "type": "string",
                      "description": "主類別"
                    },
                    "sub_category": {
                      "type": "string",
                      "description": "子類別"
                    },
                    "teacher_id": {
                      "type": "string",
                      "description": "教師ID"
                    },
                    "teacher_name": {
                      "type": "string",
                      "description": "教師名稱"
                    },
                    "teacher_avator": {
                      "type": "string",
                      "description": "教師頭像"
                    },
                    "price_quantity": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "price": {
                            "type": "number",
                            "description": "價格"
                          },
                          "quantity": {
                            "type": "number",
                            "description": "數量"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "example": {
          "status": true,
          "data": {
            "faviorites": [
              {
                "course_id": "string",
                "rate": 0,
                "name": "string",
                "content": "string",
                "main_image": "string",
                "main_category": "string",
                "sub_category": "string",
                "teacher_id": "string",
                "teacher_name": "string",
                "teacher_avator": "string",
                "price_quantity": [
                  {
                    "price": 0,
                    "quantity": 0
                  }
                ]
              }
            ]
          }
        }
      },
      "CancelFavoriteResponseModel": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "狀態"
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "訊息"
              }
            }
          }
        },
        "example": {
          "status": true,
          "data": {
            "message": "取消收藏課程成功"
          }
        }
      },
...
      "GetTeacherVideoListResponseModel": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "status": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "intro": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                },
                "video_type": {
                  "type": "string",
                  "enum": [
                    "storage",
                    "youtube"
                  ]
                },
                "course_id": {
                  "type": "string"
                },
                "teacher_id": {
                  "type": "string"
                }
              },
              "required": [
                "_id",
                "name",
                "category",
                "intro",
                "video_type",
                "teacher_id",
                "url"
              ]
            }
          }
        },
        "example": {
          "status": true,
          "data": [
            {
              "_id": "60d21b4667d0d899234d0c83",
              "name": "Sample Video 1",
              "category": "Education",
              "intro": "This is a sample video introduction.",
              "url": "https://example.com/sample-video-1",
              "video_type": "youtube",
              "course_id": "60d21b4667d0d8992e610c85",
              "teacher_id": "60d21b4667d0d8992e610c84"
            }
          ]
        }
      },
...
      "AddTeacherVideoRequestModel": {
        "type": "object",
        "required": [
          "name",
          "category",
          "intro",
          "video_type",
          "teacher_id",
          "url"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "intro": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "video_type": {
            "type": "string"
          },
          "course_id": {
            "type": "string"
          },
          "teacher_id": {
            "type": "string"
          }
        },
        "example": {
          "name": "Sample Video",
          "category": "Education",
          "intro": "This is a sample video introduction.",
          "url": "https://example.com/sample-video",
          "video_type": "youtube",
          "course_id": "60d21b4667d0d8992e610c85",
          "teacher_id": "60d21b4667d0d8992e610c84"
        }
      },
      "AddTeacherVideoResponseModel": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "status": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        },
        "example": {
          "status": true,
          "data": {
            "message": "新增影片完成"
          }
        }
      },
...
    }
  }
}